# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildTrack Construction Expense Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /auth/register:
    post:
      tags: [Auth]
      operationId: registerWithEmail
      summary: Register a local account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterWithEmailBody'
      responses:
        '200':
          description: Registered and signed in.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterWithEmailResponse'
  /auth/login/email:
    post:
      tags: [Auth]
      operationId: loginWithEmail
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginWithEmailBody'
      responses:
        '200':
          description: Signed in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginWithEmailResponse'
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and redirect to app root
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to app root.
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Session deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /crew:
    get:
      operationId: listCrew
      tags: [crew]
      summary: List crew members (optionally filter by project)
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of crew members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrewMember"
    post:
      operationId: createCrew
      tags: [crew]
      summary: Create a crew member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCrewBody"
      responses:
        "201":
          description: Created crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
  /crew/{id}:
    patch:
      operationId: updateCrew
      tags: [crew]
      summary: Update a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCrewBody"
      responses:
        "200":
          description: Updated crew member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrewMember"
    delete:
      operationId: deleteCrew
      tags: [crew]
      summary: Delete a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Crew member deleted
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    RegisterWithEmailBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        firstName:
          type: string
        lastName:
          type: string
    RegisterWithEmailResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    LoginWithEmailBody:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    LoginWithEmailResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    CrewMember:
      type: object
      required: [id, name, role, dailyRate, status, createdAt, laborCost]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string }
        rate: { type: number }
        rateType: { type: string }
        phone: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        status: { type: string }
        createdAt: { type: string }
        laborCost: { type: number }
    CreateCrewBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        role: { type: string }
        rate: { type: number }
        rateType: { type: string }
        phone: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        status: { type: string }
    UpdateCrewBody:
      type: object
      properties:
        name: { type: string }
        role: { type: string }
        rate: { type: number }
        rateType: { type: string }
        phone: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        status: { type: string }
...