# lib/api-zod/src/generated/types/createCrewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildTrack Construction Expense Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCrewBody {
  name: string;
  role?: string;
  rate?: number;
  rateType?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  projectId?: number | null;
  status?: string;
}